day 17 part 2

+++ src/main/java/com/cp/advent2022/data/common/Position2D.java
@@ -8,10 +8,10 @@ import lombok.*;
 public class Position2D {
 
     @Getter @Setter
-    private int y;
+    private long y;
 
     @Getter @Setter
-    private int x;
+    private long x;
 
     @Override
     public String toString() {
